Fix memory leak in IntersectionInfo
If there are two QSharedObject pointers to the same object but with different classes
for the template type (when the object is a derived class), the object
can be deleted if just one of the QSharedObject reference counts go to
zero. Fix this in IntersectionInfo by only returning a normal pointer to
the intersected object, and then cleanup memory leaks.