Add Vector4d constructor for Ray class
If I remember correctly, converting between Vector3d as stored in the
Ray class and Vector4d as needed for transformations with a 4x4 matrix
was being done for every intersection test.

Storing the ray origin and direction as Vector4d from the start uses a
tiny bit more memory, but saves a lot of computation.